フッターメニュー実装
android emulatorでしか確認してないけど、result画面からホーム画面に戻るボタンを押しても強制的に再スキャンされるバグがある
