+++ lib/setting.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class SettingWidget extends StatefulWidget {
+  const SettingWidget({super.key});
+
+  @override
+  State<SettingWidget> createState() => _SettingWidget();
+  
+}
+
+class _SettingWidget extends State<SettingWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: const Text('設定'),
+      ),
+      body: const Center(
+        child: Text('設定画面'),
+      ),
+    );
+  }
+}
